Some cleanups.
