+++ debug/README.md
@@ -5,21 +5,26 @@ Custom debugging tools for GDB.
 Running
 ---
 
-After you start kernel in debugging mode, launch script with -d flag
-[(detailed how to)](https://github.com/cahirwpz/mimiker#running) you can use follwing custom commands.
+After you start the kernel in debugging mode, i.e. using
+[launch](https://github.com/cahirwpz/mimiker#running) script with `-d` flag,
+you can use following custom commands:
 
-* `kdump` - Print on screen current status of kernel.
+* `kdump` - Print on screen current state of kernel structures,
 * `ktrace` - Add/remove breakpoint on selected function.
 
-kdump aguments prints:
+`kdump` parameter select which state to dump:
 
-* `free_pages` - list of free pages per segment with virtual and physical address
-* `segments` - list all memory segments with their start and end address and number of pages (currently just one)
-* `klog` - all logs currently saved in kernel
-* `threads` - all existing threads
-* `tlb` - translation lookaside buffer with addresses and flags marking if page is dirty (D) and/or global (G)
+* `free_pages` - list of free pages per segment with virtual and physical
+   addresses,
+* `segments` - list all memory segments, incl. start, end addresses and
+   number of pages (currently just one),
+* `klog` - all log messages currently saved in the kernel,
+* `threads` - all existing threads,
+* `tlb` - Translation Lookaside Buffer with addresses and flags marking if page
+  is dirty (D) and/or global (G).
 
-ktrace aguments create/remove breakpoint on:
+`ktrace` parameter selects on which predefined function create/remove
+breakpoints on:
 
-* `thread-create` - function thread_create in [thread.c](https://github.com/cahirwpz/mimiker/blob/master/sys/thread.c)
-* `ctx-switch` - function ctx_switch in [switch.S](https://github.com/cahirwpz/mimiker/blob/master/mips/switch.S)
+* `thread-create` - function [thread_create](https://github.com/cahirwpz/mimiker/blob/master/sys/thread.c)
+* `ctx-switch` - function [ctx_switch](https://github.com/cahirwpz/mimiker/blob/master/mips/switch.S)
