+++ debug/__init__.py
@@ -1,0 +1,3 @@
+import ktrace
+import kdump
+import mutex
